Added GetCompanyUsecase

+++ internal/entities/chat.go
@@ -1,8 +1,8 @@
 package entities
 
 type Chat struct {
-	Id           int    `db:"id"`
-	CompanyId    int    `db:"company_id"`
+	Id           int64  `db:"id"`
+	CompanyId    int64  `db:"company_id"`
 	TelegramId   int64  `db:"telegram_id"`
 	TelegramName string `db:"telegram_name"`
 }
